+++ backend/backend.go
@@ -15,6 +15,7 @@ import (
 	"github.com/sensu/sensu-go/backend/keepalived"
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/migration"
+	"github.com/sensu/sensu-go/backend/monitor"
 	"github.com/sensu/sensu-go/backend/pipelined"
 	"github.com/sensu/sensu-go/backend/queue"
 	"github.com/sensu/sensu-go/backend/ring"
@@ -123,8 +124,9 @@ func Initialize(config *Config) (*Backend, error) {
 
 	// Initialize eventd
 	event, err := eventd.New(eventd.Config{
-		Store: store,
-		Bus:   bus,
+		Store:          store,
+		Bus:            bus,
+		MonitorFactory: monitor.EtcdFactory(client),
 	})
 	if err != nil {
 		return nil, fmt.Errorf("error initializing %s: %s", event.Name(), err.Error())
@@ -158,8 +160,9 @@ func Initialize(config *Config) (*Backend, error) {
 	// Initialize keepalived
 	keepalive, err := keepalived.New(keepalived.Config{
 		DeregistrationHandler: config.DeregistrationHandler,
-		Bus:   bus,
-		Store: store,
+		Bus:            bus,
+		Store:          store,
+		MonitorFactory: monitor.EtcdFactory(client),
 	})
 	if err != nil {
 		return nil, fmt.Errorf("error initializing %s: %s", keepalive.Name(), err.Error())