+++ CHANGELOG.md
@@ -10,12 +10,14 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Changed
 - The Backend struct has been refactored to allow easier customization for the
 enterprise edition.
+- Use etcd monitor instead of in-memory monitor.
 
 ### Fixed
 - Fixed `sensuctl completion` help for bash and zsh.
 - Fixed a bug in build.sh where versions for Windows and Mac OS were not
 generated correctly.
 - Display the name of extensions with table formatting in sensuctl.
+- Check TTL now works with round robin checks.
 
 ## [2.0.0-beta.2] - 2018-06-28
 
@@ -63,6 +65,7 @@ or environment does not exist.
 - The profile environment & organization values are used by default when
 creating a resource with sensuctl.
 - Migrated docker image to sensu Docker Hub organization from sensuapp.
+- Use the sensu/sensu image instead of sensu/sensu-go in Docker Hub.
 
 ### Fixed
 - Prevent panic when verifying if a metric event is silenced.